Switch to preload-as-fetch per change to spec and remove prefetch support. (#9823)
Per https://github.com/w3c/preload/issues/80 we should use `fetch` instead of and empty `as` attribute value. We feature detect the change using the newly defined behavior that invalid `as` values are not reflected.

Also removes falling back to `link-rel-prefetch` as it leads to double downloads and preload itself is now very widely supported.