Percy visual diffing of AMP pages served from localhost and using a locally built runtime (#9833)
This PR adds a ruby script that starts up a local server for the AMP runtime, loads a test page from it, and grabs a snapshot via Percy. This will run on master, like before.

Visual diffing for PRs depends on making encrypted Percy tokens available on for PRs on Travis, and will come up in a separate PR.

#9469
Fixes #9928