Wait for page load before Percy snapshot (#10078)
Speculative fix that I believe should work as an implicit wait for the page to fully load.

Tested this by searching for .i-amphtml-layout, which is a CSS element that does exist on the page, and the call to has_no_css waited for a couple of seconds before returning false.

Fixes #10022