Remove development requirement for Sandbox mode (#10176)
This was initially added so that implementors wouldn't accidentally
deploy an AMP page on LaterPay's Sandbox mode, but the Sandbox payment
workflow already warns users when they are in Sandbox mode, so this is
rather unnecessary.

This also adds a warning on this overlay to let the users know they are
in development mode for our service.